Added authentification and tests

+++ part4/bloglist/controllers/blogs.js
@@ -1,15 +1,17 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 // List all Blogs
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { name : 1 })
   response.json(blogs)
 })
 
 // Show one Blog
 blogsRouter.get('/:id', async (request, response) => {
-  const blog = await Blog.findById(request.params.id)
+  const blog = await Blog.findById(request.params.id).populate('user', { name : 1 })
   if(blog) {
     response.json(blog)
   } else {
@@ -21,19 +23,53 @@ blogsRouter.get('/:id', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
 
+  // We get the User from the Token
+  const token = request.token
+  console.log(token)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+  const userObj = request.user
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.number
+    likes: body.number,
+    user: userObj._id
   })
 
   const savedBlog = await blog.save()
+
+
+  userObj.blogs = userObj.blogs.concat(savedBlog._id)
+  await userObj.save()
+
   response.status(201).json(savedBlog)
 })
 
 // Delete a Blog
 blogsRouter.delete('/:id', async (request, response) => {
+
+  const token = request.token
+  console.log(token)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+  const userObj = request.user
+
+  if(!userObj.blogs.includes(request.params.id)) {
+    return response.status(401).json({
+      error: 'L\'action est limitée au propriétaire du Blog'
+    })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
